task 5_2

+++ 5_2.py
@@ -1,9 +1,9 @@
 # Сделать калькулятор: у пользователя
 # спрашивается число, потом действие и второе
 # число
-numb1 = int(input())
+numb1 = float(input())
 action = input()
-numb2 = int(input())
+numb2 = float(input())
 if action == '+':
     print(numb1 + numb2)
 elif action == '-':
